测试方法结果可以完整写入数据库中

+++ interfaces/src/main/java/org/nix/dao/LogDao.java
@@ -38,7 +38,7 @@ public class LogDao {
             }
             statement.setDate(3,logPo.getStartTime());
             statement.setDate(4,logPo.getEndTime());
-            statement.setInt(5,logPo.getUseTime());
+            statement.setLong(5,logPo.getUseTime());
             statement.executeUpdate();
             ResultSet resultSet = statement.getGeneratedKeys();
             return getGeneratedKeys(resultSet);
